Add Keyset
A "keyset" is a set of keys for a cache, similar to how a slice is a
"view" on an array. This allows reasonably convenient and efficient
operations on multiple keys.

Doing it with a different type like this gives us the advantage that we
can reasonably easily add a "multi" version for all methods.

Fixes #4